requirements.txt
fastapi
uvicorn
pydantic
pymongo
beautifulsoup4
python-dotenv
feedparser